No longer expand prune file wildcards.
Expanding the wildcards only works if the files already exist at the time the
script is executed, which they usually don't. A possible work-around may be to
spawn a separate make process to run the prune and strip targets. That way the
wildcards will be expanded at the right time.