libs/gettext: Disable parallel make and force working iconv.
The build system does not handle the parallel make processes very well,
erroring out for instance because it tries to create a specific directory
twice.

The detection of iconv does not work reliably because it tries to execute some
code, which will obviously fail for cross-compilation. Therefore we force the
configure script into believing that all is fine.