scripts/Makefile.package: Allow for more flexible download URLs.
For one, the new FILES variable contains files that are to be downloaded, but
not extracted. Note that in contrast to the files specified by TARBALLS, the
items listed in FILES need to be *full* URLs.

An extended syntax for URLs is now possible: appending ';filename=<filename>'
to an URL will download the contents of the URL into the file <filename> in
the download directory.