Generalize and improve build-dependency handling.
The configuration variable list is now written to include/config/depends.conf
in a make compatible format and provides the "depends" make variable. This
allows simple inclusion of depends.conf in makefiles and no further processing
is required.

The platforms are now generalized in a way that subdirectories are descended
into unconditionally, just like with packages. This simplifies the code and
removes special cases.