Create dependency cache to improve performance.
This basically traverses the complete tree once and creates a list of
directories that need to be built sorted in order of dependency. This list is
written to a file that can be included by other makefiles so they no longer
need to recursively search the tree for directories to build.

The directory dependency cache is recreated whenever the platform
configuration (include/config/auto.conf) changes.