xorg/proto/xcb: Python support files are now installed in /usr/lib/python2.6.
This is actually due to the Python version on this development machine being
upgraded to 2.6. Ideally, Python should be built as part of a set of support
packages for the build host, along with perhaps Perl, glib, and other tools
required by cross-builds. Such a scheme would significantly reduce the
dependencies on the build host.