packages/libs/glib: Depend on packages/build-tools/glib.
glib needs some tools usually generated by itself during the build
process. Unfortunately it fails at correctly building these tools in
cross-compiles and relies on the build host to provide them.

To work around this, the PBS uses the packages/build-tools/glib package
to generate a glib copy for the build host in $(objtree)/build-tools
which provides the necessary tools.