packages/toolchain: Include PBS version in tools.
binutils and gcc will show this version string when called with the
--version command-line argument. It is useful in order to determine
which version of the toolchain a specific binary was built against,
using the .gnu.note.build-id ELF section.

Note that when a binary is stripped, the .gnu.note.build-id section is
also removed. Identifying the toolchain used to build a binary will
therefore only work on binaries that still have the .gnu.note.build-id
section or unstripped binaries in general.