packages/toolchain: Make language support configurable.
The list of languages supported by a toolchain can now be configured.
GCC will only build support for the selected languages, which is useful
in order to reduce the toolchain build time if Java is not required.

Note that currently the toolchain build still downloads and extracts the
Java tarball, even if support for it is not included. This is due to the
fact that the build system applies some patches against files in GCC's
Java tarball.