toolchain: Major cleanup.
This patch removes some dead code and tries to be more consistent in a
number of places. For instance, toolchain variants are no longer
supported. There used to be a multilib variant and a sysroot variant.
The only working configuration which is also actually used is the
sysroot variant, so it is simply made the default. It may be interesting
to add multilib support to it at some point in the future, however.

This patch also centralizes some variable definitions so they don't have
to repeated in every build script.

Finally some hacks which are no longer required are removed. This
concerns header files missing from the gcc or libc installation and
similar workarounds.

These changes have been tested and build all available toolchains
(except the bfin-linux-uclibc one, which really was always broken).