packages: Remove implicit makefile dependencies.
This patch removes any implicit package build step dependencies. In
order for this to work properly, all existing packages need to make
these dependencies explicit. However, this is already the case for most
packages, so this mostly consists of adding a dependency on .patch to
the .configure target, and a dependency on .configure to .build.

The majority of packages are autotools-based, so the explicit
dependencies are already given in packages/autotools{,-git}.mk.

The reason for removing the implicit dependencies is that a lot of
packages (not covered by autotools.mk) do not require a setup, patch or
configure step. With implicit dependencies, such non-required targets
always need to be specified, even if they are empty. Without implicit
dependencies, they can simply be omitted.