build system: Embed target in package filenames.
Instead of sorting the packages into subdirectories identifying the
architecture/processor and C library which the packages are built
against, package filenames now contain the complete target configuration
string, which is also used in the toolchain prefix.

In fact, categorizing by architecture/processor and C library is
actually not enough, because it doesn't account for the ABI, the
operating system and potentially other factors.