packages/core/busybox: Update to 1.17.1.
This patch also uses CONFIG_ARCH and CONFIG_CPU instead of ARCH and CPU
which are deprecated. CROSS_COMPILE should also no longer be depended
on, and $(TARGET)- should be used instead.

The fix-segfault patch can be dropped because GCC 4.5.1 (which is the
compiler in the current toolchain) now generates correct code.