Build system: Rewrite binary packaging system.
Replace the pbs-extract utility with equivalent make script.

Binary packages are now installed and built on a per-binary basis.
Essentially, the .binary target runs the build target of the new
Makefile.binary with a reference to the package's source directory.
A list of all binary packages is then constructed from the list of
.install files, and each binary package is then installed (the files
listed in the corresponding .install file are copied to a binary-
specific destination directory) and built. The .install target runs the
Makefile.binary's extract target, which extracts the binary packages
just built into the platform's sysroot directory.

The new [-q|--quiet] option of the pbs-binary and pbs-install scripts is
used to disable informational output messages from the scripts so that
the build log is cleaner.