Build system: Fix initrd and rootfs goals.
The initrd goal generates a list of tarballs based on the .install files
of selected packages and filters out all development, documentation and
localization packages. That list is then installed into a directory and
converted into an initial ramdisk (compressed cpio archive).

The rootfs does essentially the same, except that the root filesystem is
made into a squashfs image.