Build system: Introduce new version numbering.
The new version numbering scheme is time-based. It uses the year and
month as major and minor version numbers. 2010.08 would mean a release
in August 2010.

Care must be taken that the version number is bumped whenever new
changes are committed after a release. This is necessary to determine
the precise PBS version that a production image was built against.

Typically a new branch, e.g. 2010.08, would be created when 2010.08 is
released. Such a stable branch can extend the two-part version number
with a third part for patch releases. 2010.08.1 would be the first patch
release of the 2010.08 branch.

For individual platforms that branch off of a stable branch, a suffix
can be added. 2010.08-medcom would be the stable Medcom branch based on
the 2010.08 stable branch. Releases based on the platform stable branch
can use further version components. 2010.08-medcom-1 would be the first
release tagged on the stable 2010.08-medcom branch.

This should all really be scripted somehow, such that the version number
is automatically generated according to the above scheme. Another
advantage of having scripts do this would be to automatically integrate
some kind of magic into the version string if the source tree is dirty
(has uncommitted changes).