build system: Add control data to binary packages.
This patch adds basic Debian-like control data to binary packages.
Currently this includes a hook script which can be run right before the
binary package is created and a Debian package control file.

If a file named <package>.binary is found in the package source tree,
the pbs-install script will copy it into the temporariy filesystem from
which the binary package will be generated. The file will also be made
executable.

If a <package>.control file is found in the package source tree, it will
be piped through the opkg-gencontrol script, which will replace the
@TARGET@ and @VERSION@ placeholders with the correct values.

When a binary package is created, the binary script can be used to
perform various actions within the filesystem before it is actually
tarred up. Two parameters are passed to the script: the absolute path to
the temporary filesystem and the version number of the package.