toolchain: Make batch scripts executable.
Make the build-toolchains and list-toolchains batch scripts executable.
These scripts are useful when batch-building a number of toolchains. A
typical use-case would be building all available ARM-based toolchains:

	$ toolchains/list-toolchains | grep '^arm-' | \
		toolchains/build-toolchains

The build-toolchains script has a bug that causes it to abort batches
prematurely in some cases. This happens after a successful build with
more toolchains to build in the queue, as if the while-loop is broken
out of early.