build system: Generate default .config files.
If no .config file exist in the output directory, this patch generates a
default one using the KCONFIG_ALLCONFIG method[0]. Basically, this will
use an all.config file (if it exists) from the output directory and
generate a configuration based on the contained symbols.

This is mostly useful for toolchain builds. For non-toolchain builds,
this is not really meaningful and should probably be avoided. The reason
is that if no all.config file exists, a default configuration with no
symbols selected will be generated. The build process will also be
started immediately afterwards, so there is no chance to create a useful
configuration using e.g. menuconfig. Perhaps a better solution would be
to fallback to the old behaviour (error-out if no .config exists) in
cases where all.config is missing from the output directory.

[0]: http://kernel.org/doc/Documentation/kbuild/kconfig.txt