build system: Add version to top-level Makefile.
The composed version string is exported as RELEASE and can be used by
build scripts to incorporate the PBS tree version into packages or
generated files.

Most notably, the RELEASE variable should be used as release information
in the toolchain. Both binutils and gcc use this information to mark the
generated binaries.