libs/gconf: Update to 2.31.7.
In addition, this patch uses the MAJOR and MINOR variables instead of
hard-coding the major and minor version number in a number of places in
the build script. This will come in handy in future upgrades because
less of the script will need to be adapted to new version numbers.

This patch also refreshes the libtool sysroot support patch.