web/mozilla/firefox: Update to 3.6.8.
To lessen the dependency burden on the build host, this package now
build-depends on the glib and libIDL build-tools packages.

This patch also removes the version string from several directory names
installed by this package. Incorporating the version within the
directory names only makes sense if more than one version are allowed to
be installed. Since the PBS is typically used to generate small root
filesystems, multiple versions of the same browser are not supported.

To properly build against the sysroot, a patch is added which allows
passing extra compiler and linker flags.