packages: Add core/syslinux package.
This package provides the syslinux collection of bootloaders that can
boot x86 processors with images loaded from a number of filesystems,
including ext2/3/4, btrfs or FAT.

A PXE firmware is also provided in order to allow devices to boot from
the network.

This package is really only useful when building installation images
that need to install a bootloader and should not be needed in regular
"runtime" setups.