packages: Add build-tools/systemd package.
This package provides a version of systemd that runs on the build host
and is required to generate files during the build process of the
core/systemd package.

Note that this really is a flaw in systemd's build system, which should
be able to generate the files by some other, more cross-compile friendly
means.