packages: Add xorg/driver/xf86-video-vesa package.
This package provides an X.Org display driver based on the VESA
standard. Not that it has no acceleration whatsoever and should only be
used as a fallback if there is absolutely no other way to run X, in
which case it may still be better not to run X at all.