libs/libgpg-error: Update to 1.10.
This commit adds a patch that fixes an issue that occurs on some build
hosts during the generation of error code tables. The preprocessor is
invoked to extract a list of standard errno definitions, but is not
passed the correct set of preprocessor flags. When cross-compiling, the
required --sysroot flag is therefore not passed and the preprocessor may
not be able to find the correct errno.h, or not find it at all.

Oddly enough this works without a patch on Debian sid, but fails on
Fedora 12. With the new patch, both systems build the package correctly.

This commit further tries to fix the watch file, but somehow the mirror
doesn't like the kind of probing that uscan performs, so uscan still
fails to detect new upstream versions (+1 for a rewrite of uscan).