kernel/module-init-tools: Depend on core/busybox.
This is sort of hackish. The problem is that both module-init-tools and
busybox provide modprobe et al. and depending on the order in which they
are built/installed, the busybox utilities will overwrite those provided
by module-init-tools.

Making module-init-tools depend on busybox ensures that busybox will
always be installed before module-init-tools, therefore resulting in the
fully-featured utilities making it into the final root filesystem.