generic/initrd: Mount persistent storage.
If a partition named 'persist' is found, the initial ramdisk will mount
it at the /persist mountpoint within the root filesystem. This partition
is special in that it is writable and persistent across reboots and
power-cycles.

Note that this has to be done from within the initial ramdisk because
the persistent filesystem may contain data (like /etc/hostname) that is
required at the very early boot stages when not all filesystem may be
mounted yet.