packages: Add build-tools/libtool package.
This package provides the libtool library and utilities which can be
used during the build process by packages that need them. This should be
depended on by all autotools-based packages that are built from source
repositories.

Note that with this in place it may be possible to relibtoolize all
autotools-based packages during the build process in order to get rid
off all those ltmain-sysroot patches, which are no longer required when
using this libtool version.