Add uscan replacement in scripts/watch.
This python script only handles version 3 watch files and probably does
not support everything that uscan does, but it is sufficient for the
purposes of the PBS.

The old uscan infrastructure is kept in place for now, just in case.
Note that uscan is very dependent upon Debian/Ubuntu, so it won't work
on other distributions, which is the real reason for this replacement.