packages: Cleanup toolchain/glibc packages.
Use common location and version definitions for all glibc packages. Also
download bzip2 tarballs instead of xz ones because the latter don't seem
to get uploaded by default.

Update watch file to check for -ports tarballs because it is needed to
properly build on all platforms. The glibc package watch file is used as
primary, and -headers and -cleanup packages symlink to it.