packages: Add net/ifplugd package.
This packages provides the ifplugd daemon, which is used to
automatically configure or deconfigure network devices when their link
comes up or goes down, respectively.

While this package is no longer maintained upstream, it should be
evaluated if it is not a superior alternative to the very heavyweight
net/network-manager package. A rewrite would be another alternative.