net/lldpd: Add systemd support.
This commit adds systemd socket-passing support to the lldpd daemon.
This allows requests to the daemon to be accepted before the daemon is
fully started, which in turn allows the lldpd systemd service file to
not rely on some heuristic method of determining whether the control
socket has already been created or not.