Don't require prune directories and files to exist.
Some packages do fancy stuff like creating a symlink to .la files. The
prune wildcards will list both the .la file and the symlink, which will
lead to the problem that when the .la file is remove before the symlink,
the prune target will fail because the symlink is now dangling. This
patch fixes this issue by allowing files and directories that show up in
prunefiles or prunedirs respectively to not exist.