Fix building libnih with uclib.
Unlike glibc uclibc doesn't provide libintl, in such case libnih use
an internal version of libintl. However there is a bug in autoconf and
the CFLAGS are not properly set in case of an out-of-tree build.

There is also a bug in libnih install, libintl.h is included from
some installed headers, but it is not installed when building against
the internal libintl.