packages: Update util/medatom-utils to 0.1.4.
The package now builds from the official release tarball and has a watch
file to check for new versions.

A runtime dependency on the util/adak package is required. Since there
is no support for runtime dependencies within PBS, the only way to
express this is through a standard, build dependency.