packages: Remove net/network-manager package.
NetworkManager turns out not to be usable for embedded systems or
read-only filesystems. It needs a lot of patches to run properly but
still has a lot of problems like curious defaults for DHCP timeouts and
no way to configure them.

In the end, the only useful functionality provided by NetworkManager is
link monitoring, for which net/ifplugd or net/iflinkd are much better
suited.