Add support for building source distributions.
Running make O=build dist will create to source code distribution in
build/dist. It contains all the source tarball, as well as checkout
(in the svn/cvs sense ;) of the remote and local repositories used.

Note: As it use git archive to copy the pbs tree, changes in the
working copy and index WILL NOT BE INCLUDED. For local sources (ie.
src link) tar is used, so local changes will be included.