package/imlib2: Use pkg-config instead of freetype-config.
By default imlib2 is using freetype-config to determain the path for freetype.
Because the sysroot argument is not working for freetype-config, this will not
work on distobutions, like fedora, where 32 and 64 bit versions can be
installed in different versions.

So we replace the usage of freetype-config with pkg-config wich is also
provided by freetype.