buildsys: Fix depends-dirs.conf generation.
This commit gets rid of the double slashes that were previously written
when generating the depends-dirs.conf file. The double slashes are due
to the package/subdirectory name being concatenated with the parent
directory name, with the latter already terminated by a slash to mark it
as a directory. The additional separator slash is therefore not
necessary.