buildsys: Add toolchain dependencies only if available.
Toolchains are cleaned and archived when built with the build-toolchain
script in the scripts subdirectory. Therefore the depends-dirs.conf file
is not available for the toolchain and the depends-dirs list from the
platform will be used instead, resulting in duplicated targets. As a fix
this commit does not try to include the toolchain's depends-dirs.conf if
it does not exist.