packages: Update core/util-linux to 2.19.
This package includes patches that make changes to the autotools build
system files and runs autoreconf before configuring the package. As a
consequence, the ltmain-sysroot patch can be dropped because libtoolize
will be run as part of autoreconf and install recent libtool support
files. A dependency on build-tools/libtool is added to make sure that
libtool has built-in sysroot support.