buildsys: Refactor autotools build script.
This commit builds autotools-based packages completely outside of the
source tree, which is necessary to avoid polluting the source tree if
using the source symlink facility.

Furthermore, common autotools-based variables and definitions are split
into a separate file, such that packages which require the common code
but cannot use the autotools.mk script don't have to duplicate it.