buildsys: Minor cleanup.
This brings the symlink source functionality more consistent with the
rest of the build system. Namely the src symlink in $(pkgsrctree) is
renamed to source to more closely match the corresponding build
directory in $(pkgtree). Also rename the make stamp to .symlink-source
from .ln-src to make its meaning more explicit.

I known this is mainly cosmetic, but please don't hit me.