scripts/mkpasswd: Allow non-alphabetic default salt.
If a salt was not explicitly specified, the default behaviour is to use
the first two characters of the cleartext password, rotated by -1 and +1
respectively. Obviously this only works for ASCII letters (a-z and A-Z).
This commit fixes this issue by doing the rotation transformation only
for letters, using other characters as is.

BTW, it may seem seem arbitrary and not very useful to perform this
permutation of the salt but in fact I used to use a program (of which I
unfortunately don't remember the name) that behaved exactly like this to
generate password hashes in the past.