libs/libev: Fix autoreconfigure target.
This fix requires an additional patch which puts libtool's m4 files into
the m4 directory, so that aclocal will use them instead of pulling them
in from the build system.

Perhaps this is one of the issues that could be solved better by
providing the autotools (autoconf and automake packages) with the
toolchain.