xorg/app/xterm: Fix cross-compile issue.
During configure the wrong version of pkg-config is used, so the include paths
are pointing to the host machine not the sysroot. The same issues is caused
also by imake, wich also creates wrong include paths.

To fix the issue we disable imake usage and the configure were it can find the
right pkg-config tool.