core/systemd: Add patch to make system-logger.service work.
Older kernels such as 2.6.32 may not support syscall accept4, wich is a
shortcut for accept + 2x fcntl.
This patch implements a wrapper wich is used instead of accept4. It calls
accept4 and if this fails with ENOSYS it will call accept plus the two calls
to fcntl with FD_CLOEXEC and O_NONBLOCK.