libs/libxslt: Fix cross-compilation.
This commit fixes a bug caused by the libxslt build system relying on
the xml2-config script to detect the libxml2 installation. This kind of
script is known to behave badly with cross-compile builds. Instead, a
patch is added that detects libxml2 using pkg-config.

As a side-effect, the configure script needs to be regenerated. In order
to get rid of the ltmain-sysroot patch, a dependency on
build-tools/libtool is added so that the patch is no longer required.