Add a patch to fix sysroot support in libtool 2.4.
Current libtool 2.4 has a bug that is triggered in some corner cases,
usually when building a .la that depends on another .la. Within PBS this
only happens when a package builds a .la that requires another .la
within the same package (glib, gtk, gdk-pixbuf, ...) because .la files
are never installed into the sysroot.

The bug also seems to be ignored when cross-compiling for x86 on x86_64
or depending on the distribution. Fedora seems to not exhibit this
problem because it installs libraries to /usr/lib64 instead of /usr/lib
such that the linker will not find incompatible libraries in /usr/lib
when cross-compiling for a 32-bit architecture. However this is all much
too fragile.