Fix the useless recreation of source links wich break make print.
b5eb267618f39f0 introduced a dependency on the link to the source
directory. But make dereference links when getting the mtime, so it
would use the mtime from the source directory.
The rule creating the link could touch the directory to update its
mtime, but generally using the mtime of directory is always problematic.
So revert to how it was done before b5eb267618f39f0, just rely on the
stamp.