packages: Add xorg/util/util-macros package.
This package provides m4 macros that are used during the build process
of many X.Org packages. Typically they are included in release tarballs,
but some packages don't ship a properly autotoolized tarball or need to
be relibtoolized. In those cases, this package is needed to provide the
necessary macros to the autotools.