scripts: Generate more dependency information.
The depends-dirs script now generates a list of platforms and the master
platform variable in addition to the directory list.

The depends-platforms variable lists all platforms that are built. This
is currently unused but becomes more useful when platform-subclassing is
supported, allowing a platform to "inherit" a more generic one.

The master-platform variable contains the name of the master platform
(technically, the last platform in the dependency list), which is the
platform selected via Kconfig. It is useful to locate additional files
or script to be included in distribution tarballs or to be run at some
point in the build process.