Make generic platforms inheritable.
This commit introduces the concept of platform subclassing, which allows
one master platform (selected via Kconfig) to inherit from a more
generic platform by depending on it.

Note that due to the special handling, care must be taken if inheritable
platforms are conditionalized or else they may not be properly selected
and the inheritance will be broken.