core/udev: Make building extra programs optional.
This commit adds a Kconfig option that can be used to enable or disable
building extra programs. These programs have additional dependencies
that may be undesirable (e.g. for initial ramdisks).

Furthermore this commit adds a patch that allows building extras even if
libacl is not installed by disabling the udev-acl program.