multimedia/mplayer: Try to fix build properly.
The configure script shipped with mplayer seems to have some problems
with the detection of i786 CPUs. To generate the target name for the x86
architecture, the i786 is renamed to i686. For all other architectures,
the architecture name is used instead. In addition, no optimization
flags are passed to the configure script to let itself decide what the
best combination is.