packages: Update uClibc to 0.9.32.
This commit removes the accept4, epoll and pipe2 patches that have been
applied upstream. The out-of-tree patch was partially applied upstream
and the latest.patch is a collection of fixes that have been applied
upstream after the 0.9.32 release, and which are needed to build nptl
correctly.

One would think that they run test builds before releasing...