multimedia/gstreamer: Fix cross-build on non-Fedora.
This commit updates some of the libtool sysroot support patches and
fixes the cross-compilation of the packages on non-Fedora build hosts.
The problem seems to be that libtool 2.4 has a known bug when linking to
libraries in the same build tree. During relinking, it tries to link
against the corresponding libraries in /usr/lib.

The issue is fixed by the updated patches and by passing --with-sysroot
to the configure script. Note that not all the packages in this commit
immediately need the fix, but they are fixed nonetheless just in case
they may require it in the future.

Apparently this is not an issue on 64-bit Fedora, because they install
libraries to /usr/lib64, and the linker won't find incompatible
libraries.