packages: Update multimedia/vlc to 1.1.12.
This commit also adds a backported patch that adds Vout events, which
can be used to receive notifications when the number of video outputs
changes. Furthermore, support for RTSP is enabled by depending on the
multimedia/live package.

An experimental patch is added to make the OMXIL module support the
NVIDIA Tegra2 implementation.