kernel/linux: Pass custom PATH to kernel makefile.
The "env" make variable contains the PATH adjusted to include the
toolchain and the platform build-tools. These are required for the
kernel build system to find the correct tools.

This error was first triggered because the mkimage/dtc tools are
installed as platform build-tools and need to be in the PATH for the
kernel build system to find them.