multimedia/alsa/core: Fix a busy loop.
On some hardware, the ALSA library goes into a busy loop when recording
data from a capture device. The problem is that ALSA expects this to be
an atomic operation, but if additional data arrives between both atomic
barriers the complete operation is repeated. Depending on the hardware,
this may cause long delays while ALSA repeats the same loop up to 1800
times.

Note that this may not really be an ALSA library bug but could be caused
by the driver or some other component. The bug is only present during
phone calls with linphone. Looping arecord to aplay doesn't show the
same issue.