packages: Add build-tools/orc and libs/orc packages.
These packages provide the OIL runtime compiler (ORC). It is composed of
a library and a set of tools for compiling and executing very simple
programs that operate on arrays of data.

Some packages include optional support for ORC that allows certain code
paths to be accelerated. One such example is gstreamer/plugins/ffmpeg. A
follow-up patch will add optional support to this package.