package/xorg-server: Make package version depend on other selections.
Since we have binary packages for tegra2 we need fixed ABI versions of the
xserver. The only way to guarantee that we can use the binary blob, is to run
against the last version of the xserver wich provides this ABI.

This patch will specify the xserver version depending on the selected tegra
binaries.