core/busybox: Conditionally build module utilities.
If another package that provides these utilities is selected, do not
build them. The advantage is twofold: there is less code to be built and
there is no risk of the busybox package overwriting the other package's
files.

The same technique can probably be applied to other applets as well.