gstreamer/plugins/ugly: Do not disable mad plugin.
In the past we disable mad plugin, to enshure we use ffdec_mp3 instead of mad.
This was nessessary because the mad plugin cause trouble with A/V sync. Now
that we have a better solution for this special behavior and since mad is
required to make webradio work, we can enable it by default. For more details
see remote-control commit 32073f7e19fd3d985655d1dc7a6ce0214df7e969.