packages: Update core/udev to 180.
This commit also adds a dependency on kernel/kmod, which is now used by
udev to load modules instead of relying on module-init-tools. Also,
starting with this commit all files are now installed below /usr in
preparation for the /usr merge. However, symlinks are provided in the
old locations for backwards-compatibility.