packages: Update core/systemd to 39.
This commit also adds a dependency on libs/xz which is used by systemd
to compress the journal. Note that systemd's journal now takes over the
/dev/log socket and any syslog daemons should be monitoring the socket
in /run/systemd/journal/syslog instead.

Furthermore, most files are now installed below /usr in preparation for
the /usr merge, but symlinks are created in the old locations for
backwards compatibility.