libs/glib-networking: Fix fix-null-deref patch.
The previous patch was actually wrong because it was asserting that the
connection when performing a handshake was always a server connection.
This obviously breaks handshaking for client connections.

Instead of checking for a valid server TLS connection, we check for a
valid TLS connection and keep the remaining code as-is.