build: Rebuild packages if their makefile changes.
If a package's makefile changes, it should be rebuilt. This restores the
behaviour that packages are automatically rebuilt if the version number
changes. In fact it triggers rebuild much more often, for example if the
configure options change or files are added to the list of files to be
stripped. Each time the package's makefile changes, its SHA1 digest is
computed and compared to the SHA1 digest stored in the package's hash
file. If they don't match, the new digest is written to the hash file
and the package will be rebuilt.

Packages which have some of their package information split off into a
separate file (e.g. Makefile.location) can list those extra dependencies
in the optional per-package depends.mk file. The latter will be included
by the main depends.mk file and is also part of the digest to make sure
that any changes to the depends.mk file will also cause a the digest to
change and therefore trigger a rebuild. Per-package depends.mk files may
use the special $(dep-hash-file) and $(dep-obj) variables to substitute
the hash file and the package directory respectively.

Example:

	$(dep-hash-file): $(dep-obj)/Makefile.location

Checking for changes based on the SHA1 digest avoids needless rebuilds
of a package when only the modification time of the files changes, which
can be caused by source control operations.

Note that this commit will trigger a complete rebuild of *all* packages
because of the missing hash files.