packages: Add build-tools/expat package.
Building this package as part of the build-tools is necessary to
properly build the build-tools/librpc package, which depends on a patch
to expat (adding a pkg-config file) that is not upstream yet.

This commit also splits off the version information of the libs/expat
package into a separate file so that it can be reused by this new
package.

Furthermore, a common makefile snippet is provided to allow autotools-
based build-tools packages to be auto-reconfigured before building (as
required by the pkg-config patch).