Fix package hash file check.
The hash file check command uses one excess sha1sum invocation. The SHA1
digest is computed once for each hash file dependency and the resulting
digest file is then checked against the existing hash file. If they
don't match, the SHA1 digest of the SHA1 digest file is stored in the
hash file.

I don't know if anyone understand this. In this case the code is clearer
than the words. Then again, I didn't get the code right the first time,
so maybe it isn't so clear after all.