packages: Add toolchain/ccache package.
This package provides the ccache compiler cache that can be used to
speed up recompilation. It works by caching the results of a build and
looking them up in the cache in subsequent builds to reduce the build
time.

Furthermore, the gcc wrapper script is modified such that it can use
ccache if the CCACHE environment variable is set. If so, it calls the
command defined by CCACHE followed by the regular command line. Setting
the CCACHE environment variable to ccache makes the build automatically
use it.