Introduce the CCACHE make variable.
This commit allows setting the CCACHE variable on the command line when
starting a build. Setting it to ccache will cause the build to export
the CCACHE environment variable during package builds, which in turn
will cause the gcc wrapper script to call the cross-compiler under the
control of ccache:

	$ make O=build CCACHE=ccache

The cache directory will be located at the top of the PBS source tree.

Note that if any value other than ccache is specified, the wrapper will
blindly execute the cross-compiler under the control of that command
instead.