medatom/initrd: Add default configuration.
Building this specific configuration can be done as follows:

	$ make ARCH=x86 O=build/medatom-initrd medatom-initrd_defconfig
	$ make O=build/medatom-initrd

Currently it is still required to specify ARCH=x86 for the scripts to
find the correct default configuration. The reason is that the Kbuild
system is taken from the Linux kernel, which splits up the default
configurations this way. For PBS perhaps a shared directory for all
default configurations may be a better choice. Then again, separating by
architecture has its advantages as well (mainly that the directory is
less cluttered).