generic/initrd: Move /dev to new root filesystem.
Moving the /dev filesystem to the new root filesystem prevents a strange
error by systemd, which complains about it not being able to open the
/dev/kmsg device for logging. systemd still goes and mounts a devtmpfs
on /dev later, so the error is mostly harmless. It doesn't hurt to keep
the old /dev, though, and should even provide for some minimal speed up
of the boot process.