core/syslinux: Make build resilient against environment.
Under some circumstances, the syslinux build system will fail miserably.
For instance, if the top-level PBS make is run with the ARCH variable
set via the command-line, the build will fail. This can be solved by
passing an empty MAKEFLAGS variable to the syslinux build system, which
will cause ARCH to be unset and the build succeeding.