Make Linux kernel build scripts more consistent.
This commit adds a new package, kernel/linux, which builds the Linux
kernel in a way that is more consistent with other packages. Several
ways of specifying the source are now available:

  * tarball: A Linux kernel tarball is downloaded and extracted before
      running the build. This is really only useful if a all drivers and
      patches have been mainlined or if a vendor provides a custom
      tarball. Note that this mode assumes that the tarball extracts
      into a directory named after its basename (i.e. linux-3.2.5.tar.xz
      extracts to linux-3.2.5).

  * git checkout: A Linux kernel repository is cloned on the build host
      and a specified branch or tag is checked out before running the
      build.

  * local directory: The Linux kernel is built from a local directory
      that is typically the result of extracting a tarball or cloning a
      Linux kernel git repository.

During development, the recommended mode is the last of these three
because it allows changes to be immediately visible in the build without
having to commit and push to a git repository or creating a tarball. A
very good workflow is by pointing the local directory to a git checkout
from where tested changes can be committed and pushed.

Release builds should be based on either of the first methods. If all
code required for a build is in a released upstream kernel, the
corresponding tarball can be used with the first mode. The second mode
is ideal if a project is under active development and not all code has
been mainlined because it allows to build e.g. from a release tag in a
downstream repository.

This commit also removes the obsolete kernel/linux.git package and
updates the toolchain/linux-headers package to be consistent with the
new kernel/linux package. The old kernel/linux.git package is dropped
and all configuration options are updated to use the new package.

Also, two new default configurations are added for the CSC and PTIP
Classic projects. They will need to be updated with the correct local
directories though. Note that the updated default configurations remain
compatible in that they still build from the same local directories.
Eventually all default configurations should be modified to point to the
canonical upstream repository, such that when a default configuration is
built, the last known good version of the Linux kernel for the platform
will be used. The local directory mode should only be used locally for
development in order to avoid embedding knowledge of the developer's
directory hierarchy within the default configuration.